PM-443 moving all tests to same folder, specifying what they are testing
