+++ src/routes/marketlist/store/test/builder/oneMarket.js
@@ -1,0 +1,66 @@
+/* eslint-disable */
+export default {
+  "count": 3,
+  "next": null,
+  "previous": null,
+  "results": [
+    {
+      "contract": {
+        "creationBlock": 4818543,
+        "creator": "0x9eab578556de5782445ec036f25a41902ba19eeb",
+        "factoryAddress": "0xdde21f8be9e50b9b1165ee302e6b468cf30e4c4c",
+        "creationDate": "2017-12-29T13:39:44",
+        "address": "0xa625c9ccf5860e9709a31cbac1ffd6fc557558f6"
+      },
+      "event": {
+        "contract": {
+          "creationBlock": 4818540,
+          "creator": "0x9eab578556de5782445ec036f25a41902ba19eeb",
+          "factoryAddress": "0x5b21ba38a0db91cc43aa7daba778979758e67991",
+          "creationDate": "2017-12-29T13:39:29",
+          "address": "0x2077708a049f059ee7a7878071b5adca92e6a4ba"
+        },
+        "collateralToken": "c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2",
+        "oracle": {
+          "contract": {
+            "creationBlock": 4818537,
+            "creator": "0x9eab578556de5782445ec036f25a41902ba19eeb",
+            "factoryAddress": "0xc4f4b8c0259c8264376fc32984e2a8cf4a3c70ca",
+            "creationDate": "2017-12-29T13:39:02",
+            "address": "0x02d7f617bc27b2a09675d2c095c1d0ffc761afd5"
+          },
+          "isOutcomeSet": false,
+          "owner": "0x9eab578556de5782445ec036f25a41902ba19eeb",
+          "eventDescription": {
+            "ipfsHash": "QmfGcwGtG7LoPSD3wQk3hNNKZkRAszAgrzeS6GrB7Us9np",
+            "description": "How much of the total Ethereum transactions in the last 1500 blocks will be made by CryptoKitties game actions in the Ethereum network? The winning outcome will be calculated according to the sum of interactions with the smart contract address: 0x06012c8cf97bead5deae237070f9587f8e7a266d (CryptoKitties) and the address: 0xb1690c08e213a35ed9bab7b318de14420fb57d8c (CryptoKitties Auction) which make up for all the interactions made with CryptoKitties contracts. Source: https://ethgasstation.info/gasguzzlers.php",
+            "title": "How much will Cryptokitties transactions make up of the entire Ethereum network transactions by December 29th, in the last 1500 blocks?",
+            "resolutionDate": "2017-12-30T00:00:00",
+            "decimals": 0,
+            "unit": "%"
+          },
+          "type": "CENTRALIZED"
+        },
+        "isWinningOutcomeSet": false,
+        "lowerBound": "1",
+        "upperBound": "20",
+        "type": "SCALAR"
+      },
+      "marketMaker": "7ee6e9dc512b0fe5cb5e28697cfc34375a9adc4b",
+      "fee": 0,
+      "funding": "10000000000000000",
+      "netOutcomeTokensSold": [
+        "15211146333519251",
+        "0"
+      ],
+      "stage": 1,
+      "tradingVolume": "9523809523809680",
+      "withdrawnFees": "0",
+      "collectedFees": "0",
+      "marginalPrices": [
+        "0.7416",
+        "0.2584"
+      ]
+    }
+  ]
+}
