+++ src/routes/marketlist/store/test/openMarkets.selector.spec.js
@@ -2,63 +2,65 @@ import moment from 'moment'
 import { List } from 'immutable'
 import { MARKET_STAGES } from '../models/market'
 import { REDUCER_ID } from '../reducers/market'
-import aMarket from '../reducers/market.builder'
-import { marketSelector, openMarketSelector } from './index'
+import { openMarketSelector } from '../selectors'
+import aMarket from './builder/index.builder'
 
 const openMarketTests = () => {
-  it('get empty immutable List when they are not loaded', () => {
-    // GIVEN
-    const emptyList = List([])
-    const reduxStore = { [REDUCER_ID]: emptyList }
+  describe('Market List Selector[openMarketSelector]', () => {
+    it('should return 2 open markets if there backend only have two of three opened', () => {
+      // GIVEN
+      const aClosedMarket = aMarket()
+        .withDate(moment().subtract(1, 'M'))
+        .withStage(MARKET_STAGES.MARKET_CLOSED)
+        .get()
 
-    // WHEN
-    const marketListState = marketSelector(reduxStore)
+      const aOpenMarketViaStage = aMarket()
+        .withStage(MARKET_STAGES.MARKET_FUNDED)
+        .get()
 
-    // THEN
-    expect(marketListState).toEqual(emptyList)
-  })
+      const aOpenMarketViaDate = aMarket()
+        .withDate(moment().add(1, 'M'))
+        .withStage(1)
+        .get()
 
-  it('return 2 open markets if there backend only have two of three opened', () => {
-    // GIVEN
-    const aClosedMarket = aMarket()
-      .withDate(moment().subtract(1, 'M'))
-      .withStage(MARKET_STAGES.MARKET_CLOSED)
-      .get()
+      const markets = List([aClosedMarket, aOpenMarketViaStage, aOpenMarketViaDate])
+      const reduxStore = { [REDUCER_ID]: markets }
 
-    const aOpenMarketViaStage = aMarket()
-      .withStage(MARKET_STAGES.MARKET_FUNDED)
-      .get()
+      // WHEN
+      const openMarkets = openMarketSelector(reduxStore)
 
-    const aOpenMarketViaDate = aMarket()
-      .withDate(moment().add(1, 'M'))
-      .withStage(1)
-      .get()
+      // THEN
+      expect(2).toEqual(openMarkets)
+    })
 
-    const markets = List([aClosedMarket, aOpenMarketViaStage, aOpenMarketViaDate])
-    const reduxStore = { [REDUCER_ID]: markets }
+    it('should return 0 open markets if there is there is one market but no open', () => {
+      // GIVEN
+      const aClosedMarket = aMarket()
+        .withDate(moment().subtract(1, 'M'))
+        .withStage(MARKET_STAGES.MARKET_CLOSED)
+        .get()
 
-    // WHEN
-    const openMarkets = openMarketSelector(reduxStore)
+      const markets = List([aClosedMarket])
+      const reduxStore = { [REDUCER_ID]: markets }
 
-    // THEN
-    expect(2).toEqual(openMarkets)
-  })
+      // WHEN
+      const openMarkets = openMarketSelector(reduxStore)
 
-  it('return 0 open markets if there is no open markets', () => {
-    // GIVEN
-    const aClosedMarket = aMarket()
-      .withDate(moment().subtract(1, 'M'))
-      .withStage(MARKET_STAGES.MARKET_CLOSED)
-      .get()
+      // THEN
+      expect(0).toEqual(openMarkets)
+    })
 
-    const markets = List([aClosedMarket])
-    const reduxStore = { [REDUCER_ID]: markets }
+    it('should return 0 open markets if there is no open markets loaded in store', () => {
+      // GIVEN
+      const markets = List([])
+      const reduxStore = { [REDUCER_ID]: markets }
 
-    // WHEN
-    const openMarkets = openMarketSelector(reduxStore)
+      // WHEN
+      const openMarkets = openMarketSelector(reduxStore)
 
-    // THEN
-    expect(0).toEqual(openMarkets)
+      // THEN
+      expect(0).toEqual(openMarkets)
+    })
   })
 }
